Get last activity when starting monitoring server
This implements getting the last activity of the garage door when the
monitoring server is started. Allowingthe  monitoring server to be up to
date immediately without waiting for a change state in the garage door
to update it's current state global. Ultimately having the
change state readilt available to be read by services at all times.