SG-22: Make json and base index if none exists
This fix adds a json file with the "garagedoor_usage_log" index if the
json does not exist. This allows bypassing an error of no json file
exists. If no json log entries exist in the index, an "N/A" will be
added in the HTTP response as a placeholder.