SG-35: Implement backend HTTPS.
In order for google chrome to allow notifications, the frontend needs to
be HTTPS. If the frontend is HTTPS then it can only makes HTTPS requests
to the backend. This commit enables only allowing HTTPS requests. This
also gives the ability to encrypt the username and password being sent
to the garage control server.