Empty Json logging errors fix
This fix is for the issue of the json not being created when it doesn't
exist this causing the application to error out and crash. Also, if no
entry exists in the json log, the logging server will give out a 500
error instead of N/A before this fix. That issue is also fixed.